Shocking that the cohort select still works
